Added Preference page and few categories
I have added the preference page in the Settings. When the user goes into the Settings page and clicks on "Discovery" he will have the option to choose his favourite categories.

For now I have added "Arts" and "Comedy". The only problem is that when the user selects new categories or deletes categories, the discovery page is not updated right away.

I also have to add all the other categories and a cover page in the discovery page for the eventuality that he never configured his "Preferences" in Settings.